ENGR00215945-3: caam: Improve error recovery on failed RNG4 kickstart
RNG4 requires a kickstart process to transition into running mode.
In the case that this kickstart process errors, the driver is shut
back down (under the assumption that internal random padding of keys
or data cannot occur).

In an isolated case, the kickstart failed to start the RNG, an error
was returned, and the driver attempted to de-register an RNG function
that never completed, causing a crash. (This is difficult to test for
without manual intervention).

Therefore, amended the driver shutdown process to only de-register
the hardware RNG when an instance kickstarted without error.

This does NOT correct the kickstart problem, only the consequences.

Signed-off-by: Steve Cornelius <steve.cornelius@freescale.com>
Signed-off-by: Terry Lv <r65388@freescale.com>