ARM: i.MX6: cm-fx6: fix the UART clock name
In function cm_fx6_timer_init(), prior to early_console_setup(), there
is a call to clk_get_sys() to get the clock for the early console setup.
The clock name is wrong: cm-fx6 uses UART4 and not UART1 for the
console.
Nevertheless, this has worked so far, because both UART1 and UART4 are
linked to the same clock structure.
So, this fix does not change any functionality, but only assures the
clk_get_sys() function is called with the correct clock name.

Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>