ARM: i.MX6: prevent double regulators init
The function mx6_cpu_regulator_init() is called from both the
mxc_cpufreq_driver_init() function and the cm_fx6_init().
As a result, the cpufreq driver failes to scale the voltage and we have
multiple messages in kernel log:
COULD NOT SET GP VOLTAGE!!!!
COULD NOT SET SOC VOLTAGE BACK!!!!

Prevent this by restricting the mx6_cpu_regulator_init() call to only
one time.
Currently, there is a common way to solve the problem and remove the
call to mx6_cpu_regulator_init() from the board specific function, but
leave it in the mxc_cpufreq_driver_init().

We are not sure that the common way is the best one as we'd like to see
those regulators initialized as soon as possible in the boot process.
So calling the mx6_cpu_regulator_init() from cm_fx6_init() is much
earlier than doing it from mxc_cpufreq_driver_init() as it is a
module_init() call.

Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>