video: mxc_hdmi: enable non-CEA video modes support
To be more precise on proper support of non-CEA video modes,
we reject those non-CEA modes that have a CEA match.
This is done in order to eliminate selection of unsupported
modes when there are just as good supported modes.

According to Freescale, they do not support non-CEA modes.
According to BoundaryDevices, this restriction does not appear necessary.
According to our observations, this fix quite increases the amount of
EDID modes that are accepted by i.MX6 graphics.

Reference:
[1] https://community.freescale.com/thread/305063
[2] http://boundarydevices.com/hdmi-cea-mode/

Signed-off-by: Andrey Gelman <andrey.gelman@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>