Revert "ARM: i.MX6: cpufreq: supress excessive messaging"
The commit in subject moved the error messages regarding the
regulator framework from error level to debug.
The error messages were real (and not debug) and pointed to a problem in
the platform regulators handling.

The problem has been fixed by commits:
"ARM: i.MX6: prevent double regulators init"
and
"ARM: i.MX6: cm-fx6: fix crashes related to DVFS",
and the messages are not appearing anymore, so revert the commit.

This reverts commit 07cf9f2a7fef2b864e5f5ba8053653bfed4193cf.

Signed-off-by: Valentin Raevsky <valentin@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>