Revert "ARM: i.MX6: cm-fx6: suppress annoying debug messages"
The commit in subject commented out the error messages regarding the
regulator framework.
The error messages were real (and not debug) and pointed to a problem in
the platform regulators handling.

The problem has been fixed by commits:
"ARM: i.MX6: prevent double regulators init"
and
"ARM: i.MX6: cm-fx6: fix crashes related to DVFS",
and the messages are not appearing anymore, so revert the commit.

This reverts commit 3011a87407953fe5f03f1b78c0401a6dd1beeb56.

Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>