ARM: i.MX6: cm-fx6: fix SD3 mux on Solo SoCs
The baseboard_sd3_init() function initialized the sd3_pads and
sd3_pads_cnt variables prior to checking the SoC variant and thus the
pads were always initialized as for Quad/Dual SoC variant.
This is incorrect for the Solo variant and may lead to instabilities.
Fix the baseboard_sd3_init() function logic to initialize the sd3_pads
after the SoC variant is known.

Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>